requirements.txt
accelerate==0.23.0
addict==2.4.0
aiofiles==23.2.1
aiohttp==3.8.5
aiosignal==1.3.1
aliyun-python-sdk-core==2.14.0
aliyun-python-sdk-kms==2.16.2
altair==5.1.1
annotated-types==0.5.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.11.2
black==23.7.0
blis==0.7.11
catalogue==2.0.10
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==2.1.1
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33
cloudpathlib==0.16.0
cmake==3.25.0
colorama==0.4.6
colorlog==6.7.0
confection==0.1.3
contextualSpellCheck==0.4.4
contourpy==1.1.1
crcmod==1.7
cryptography==41.0.5
curated-tokenizers==0.0.8
curated-transformers==0.1.1
cycler==0.11.0
cymem==2.0.8
decorator==4.4.2
distlib==0.3.6
editdistance==0.6.2
einops==0.6.1
exceptiongroup==1.1.3
executing==1.2.0
fastapi==0.103.1
ffmpy==0.3.1
filelock==3.9.0
flake8==6.1.0
fonttools==4.42.1
freetype-py==2.4.0
frozenlist==1.4.0
fsspec==2023.9.0
ftfy==6.1.1
gdown==4.6.4
glcontext==2.5.0
gradio==4.13.0
gradio_client==0.8.0
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
huggingface-hub==0.20.2
hydra-colorlog==1.2.0
hydra-core==1.3.2
idna==3.4
imageio==2.31.5
imageio-ffmpeg==0.4.9
imgui==2.0.0
importlib-metadata==6.8.0
importlib-resources==6.1.0
ipdb==0.13.13
ipython==8.15.0
jedi==0.19.0
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.3.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
langcodes==3.3.0
lazy_loader==0.3
lightning-utilities==0.9.0
lit==15.0.7
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.8.0
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
mmengine==0.9.1
model-index==0.1.11
moderngl==5.8.2
moderngl-window==2.4.5
moviepy==1.0.3
mpmath==1.2.1
multidict==6.0.4
multipledispatch==1.0.0
murmurhash==1.0.10
mypy-extensions==1.0.0
networkx==3.0
nltk==3.8.1
numpy==1.24.1
omegaconf==2.3.0
opencv-contrib-python-headless==4.8.1.78
opencv-python==4.8.1.78
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.0.28
ordered-set==4.1.0
orjson==3.9.7
oss2==2.17.0
packaging==23.1
pandas==2.1.1
parso==0.8.3
pathspec==0.11.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.11.0
preshed==3.0.9
proglog==0.1.10
prompt-toolkit==3.0.39
protobuf==4.24.4
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.11.0
pycparser==2.21
pycryptodome==3.19.0
pydantic==2.3.0
pydantic_core==2.6.3
pydub==0.25.1
pyflakes==3.1.0
pyglet==2.0.9
Pygments==2.16.1
PyOpenGL==3.1.0
pyparsing==3.1.1
pyrender==0.1.45
pyrr==0.10.3
PySocks==1.7.1
python-dateutil==2.8.2
python-multipart==0.0.6
pytorch-lightning==2.0.9
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
regex==2023.8.8
requests==2.28.2
roma==1.4.1
rpds-py==0.10.3
safetensors==0.3.3
scikit-image==0.22.0
scikit-learn==1.3.1
scikit-video==1.1.11
scipy==1.11.2
semantic-version==2.10.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
shellingham==1.5.4
six==1.16.0
smart-open==6.4.0
smplx==0.1.28
sniffio==1.3.0
soupsieve==2.4
srsly==2.4.8
stack-data==0.6.2
starlette==0.27.0
sympy==1.11.1
tabulate==0.9.0
termcolor==2.3.0
thinc==8.2.1
threadpoolctl==3.2.0
tifffile==2023.9.26
tokenizers==0.13.3
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.0
torchmetrics==1.1.2
tqdm==4.65.2
traitlets==5.10.0
transformers==4.33.2
trimesh==3.23.5
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.13
uvicorn==0.23.2
wasabi==1.1.2
wcwidth==0.2.6
weasel==0.3.3
websockets==11.0.3
yapf==0.40.2
yarl==1.9.2
zipp==3.17.0
